requirements.txt
aliyun-log-python-sdk==0.6.31
aliyun-python-sdk-core-v3==2.8.9
aliyun-python-sdk-ecs==4.9.5
aliyun-python-sdk-vpc==3.0.2
behave==1.2.6
certifi==2018.4.16
chardet==3.0.4
elasticsearch==6.3.0
enum34==1.1.6
idna==2.7
parse==1.8.4
parse-type==0.4.2
protobuf==3.6.0
PyHamcrest==1.9.0
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
